### 2019-04-28 #1416 @whedon
[PRE REVIEW]: qtl2pleio: Testing pleiotropy vs. separate QTL in multiparental populations
**Submitting author:** @fboehm (<a href="http://orcid.org/0000-0002-1644-5931">Frederick Boehm</a>)
**Repository:** <a href="https://github.com/fboehm/qtl2pleio" target ="_blank">https://github.com/fboehm/qtl2pleio</a>
**Version:** v0.1.3
