### 2019-05-02 c549ae3
updated install.R to latest version of qtl2pleio
