### 2019-05-02 cbbbec2
updated README to facilitate website creation with pkgdown
