### 2019-05-02 #1416 @csoneson
👋@sirsharpest - would you be interested in reviewing this submission for JOSS? Looks like a great fit to your expertise.
