### 2019-05-02 #1416 @csoneson
@fboehm - thanks for your suggestion. I noticed that @pjotrp has relatively recently published together with one of the authors of this submission, which would not be compatible with the JOSS conflict of interest statement. I have invited a couple of other great people - let's see if they are available!
