### 2019-05-05 #1416 @fboehm
Thanks so much, @csoneson and @vjcitn! @csoneson, is there a list of people who are certified to serve as reviewers? One person who might be willing to review is Robert Corty of U. North Carolina. I don't know if he has reviewed for JOSS yet.
