### 2019-05-06 #1416 @csoneson
@fboehm they should get a notification by being mentioned in this issue. @nhejazi, @rmflight, @skummerf, @dpolychr (and ping @SirSharpest, @rcorty) - would any of you be available and interested in reviewing this submission for JOSS?
