### 2019-05-06 #1416 @dpolychr
Thanks for the invitation but QTL testing is not my expertise either. Good luck with the submission and I would be happy to help out with requests that fall within my area of expertise
