### 2019-05-06 f920697
added to Dockerfile call to devtools::install_dev_deps()
