+++ Dockerfile
@@ -1,2 +1,4 @@
 FROM rocker/verse:3.5.3
 
+RUN R -e 'devtools::install_dev_deps()'
+
