+++ Dockerfile
@@ -1,4 +1,4 @@
 FROM rocker/verse:3.5.3
 
-RUN R -e 'devtools::install_dev_deps()'
+RUN R -e 'devtools::install_dev_deps(".")'
 
