### 2019-05-06 9662edb
fixed Dockerfile per https://github.com/karthik/rstudio2019/blob/master/binder-notes.md
