### 2019-05-06 d273e9c
added comma to remotes field
