+++ DESCRIPTION
@@ -51,7 +51,7 @@ LinkingTo:
 VignetteBuilder:
     knitr
 Remotes:
-    fboehm/gemma2
+    fboehm/gemma2,
     rqtl/qtl2
 Encoding: UTF-8
 Language: en-US
