### 2019-05-07 #1416 @rcorty
I’d like to consider reviewing, but I can’t see the pdf on mobile.  It says the file is corrupted?
