### 2019-05-07 #1416 @csoneson
Thanks @rcorty - I just checked and it works fine for me, both on the laptop and on the mobile (if I choose 'Desktop version' for GitHub on the mobile). I just pulled down the paper and sent it to you by email, if that's helpful to have a first look.
