### 2019-05-07 #1416 @csoneson
@vjcitn, @rcorty - thanks for agreeing to review this JOSS submission! Please head over to issue #1435 for the actual review.
