### 2019-05-08 #1435 @rcorty
license file doesn't seem detailed enough
