### 2019-05-08 #1435 @rcorty
In the summary, the authors write "For a pair of traits that map to a single genomic region, they formulated the test with the null hypothesis being pleiotropy (the two traits are affected by a single QTL) against the alternative hypothesis of two separate QTL, with each affecting exactly one trait in the pair."

QTL = quantitative trait locus, and it seems like there is only one locus being discussed.  Would it be clearer to say "...the alternative hypothesis of distinct causal variants..."
