### 2019-05-08 #1435 @rcorty
To make it easy for people to go through your README, consider making `pr` available somehow, so users don't have to computer that -- it takes a while and that computation is not relevant for `qtl2pleio`.
