### 2019-05-08 #1435 @csoneson
@rcorty Have a look here for the expected format of a JOSS submission (including the paper itself): https://joss.readthedocs.io/en/latest/review_criteria.html
