### 2019-05-08 a0b556c
updated README.Rmd and main vignette to not mention qtl2convert, since we don't use it
