+++ README.Rmd
@@ -57,12 +57,11 @@ install.packages("devtools")
 devtools::install_github("fboehm/qtl2pleio")
 ```
 
-You may also wish to install [R/qtl2](https://kbroman.org/qtl2) and
-the [`qtl2convert`](https://github.com/rqtl/qtl2convert) package. We
-will use both below.
+You may also wish to install the [R/qtl2](https://kbroman.org/qtl2). We
+will use it below.
 
-```{r install-qtl2convert, eval = TRUE}
-install.packages(c("qtl2", "qtl2convert"), repos="http://rqtl.org/qtl2cran")
+```{r install-qtl2, eval = TRUE}
+install.packages("qtl2", repos="http://rqtl.org/qtl2cran")
 ```
 
 
@@ -71,12 +70,11 @@ install.packages(c("qtl2", "qtl2convert"), repos="http://rqtl.org/qtl2cran")
 Below, we walk through an example analysis with Diversity Outbred
 mouse data. We need a number of preliminary steps before we can
 perform our test of pleiotropy vs. separate QTL. Many procedures rely
-on the R package `qtl2`. We first load the `qtl2`, `qtl2convert`,
+on the R package `qtl2`. We first load the `qtl2`
 and `qtl2pleio` packages.
 
 ```{r pkgs}
 library(qtl2)
-library(qtl2convert)
 library(qtl2pleio)
 ```
 
@@ -88,7 +86,7 @@ We'll consider the
 the [`qtl2data`](https://github.com/rqtl/qtl2data) repository.
 We'll download the DOex.zip file before calculating founder allele dosages.
 
-```{r download-allele-probs}
+```{r download-DOex}
 file <- paste0("https://raw.githubusercontent.com/rqtl/",
                "qtl2data/master/DOex/DOex.zip")
 DOex <- read_cross2(file)
