### 2019-05-08 7fdddd6
called usethis::use_mit_license to add LICENSE.md and LICENSE and to add LICENSE.md to .Rbuildignore (for CRAN)
