+++ DESCRIPTION
@@ -3,12 +3,12 @@ Package: qtl2pleio
 Title: Testing pleiotropy vs. separate QTL in multiparental populations
 Version: 0.1.3
 Authors@R:
-    person(given = "Frederick",
+    person(given = "Frederick J.",
            family = "Boehm",
            role = c("aut", "cre"),
            email = "frederick.boehm@gmail.com",
            comment = c(ORCID = "0000-0002-1644-5931"))
-Maintainer: Frederick Boehm <frederick.boehm@gmail.com>
+Maintainer: Frederick J. Boehm <frederick.boehm@gmail.com>
 Description: The package contains code that implements an
     adaptation of Jiang & Zeng's (1995) likelihood ratio test for testing
     the null hypothesis of pleiotropy against the alternative hypothesis,
