### 2019-05-08 79bec26
updated DESCRIPTION by removing periods after middle initial & adding cph to my name
