### 2019-05-11 0f81453
removed R code to install qtl2pleio in install.R file
