+++ install.R
@@ -12,4 +12,3 @@ cran_pkgs<- c("MASS",
 install.packages(cran_pkgs)
 devtools::install_github("rqtl/qtl2@b5785d5")
 devtools::install_github("fboehm/gemma2@2872396")
-devtools::install_github("fboehm/qtl2pleio")
