### 2019-05-11 e8cc81b
stopped using DESCRIPTION with mybinder & started using install.R
