+++ Dockerfile
@@ -6,5 +6,4 @@ RUN chown -R ${NB_USER} ${HOME}
 
 ## Become normal user again
 USER ${NB_USER}
-RUN wget https://github.com/fboehm/qtl2pleio/raw/master/install.R && \
-R -e 'source("install.R")'
+RUN R -e 'source("install.R")'
