### 2019-05-13 998bc03
updated install.R to use versions R package
