+++ install.R
@@ -9,6 +9,7 @@ cran_pkgs<- c("MASS",
 "broman",
 "devtools",
 "rmarkdown")
-install.packages(cran_pkgs)
+install.packages("versions")
+versions::install.dates(cran_pkgs, dates = "2019-03-01")
 devtools::install_github("rqtl/qtl2@b5785d5")
 devtools::install_github("fboehm/gemma2@2872396")
