### 2019-05-16 #1435 @vjcitn
I found it easy to install the package and run two of the vignettes.  I did not attempt the "bootstrap cluster" vignette, which has references to HTCondor and SQUID ... it seems informative for users in a specific sort of environment.  The provision of bootstrap realizations from the legacy analyses is helpful for readers aiming to understand the significance procedure.

I think it would be helpful to create print methods for the S3 classes that typically have large volumes of information associated with them.  Print methods that summarize structure and aspects of content prevent long screendumps when users want to poke into an intermediate calculation.

I have little knowledge of C++.  However I find it strange that this package src/ needs to define C++ methods for matrix x matrix and matrix x vector.  These operations are not already available in standard libraries?

Package build is slow, so I attempted to run the unit tests by hand, and ran into

```
> test_that("find_lin_indep_cols works", {
+
+   set.seed(20151130)
+   x <- cbind(1, sample(0:1, 200, repl=TRUE))
+
+   expect_equal(sort(find_lin_ .... [TRUNCATED]
Error: Test failed: 'find_lin_indep_cols works'
* could not find function "find_lin_indep_cols"
1: expect_equal(sort(find_lin_indep_cols(x)), c(1, 2)) at test-drop_depcols.R:12
2: quasi_label(enquo(object), label, arg = "object")
3: eval_bare(get_expr(quo), get_env(quo))
4: sort(find_lin_indep_cols(x))
```

By and large I consider this a fine OSS project and consider the review complete.  Let me know if
more information is needed.
