### 2019-05-22 #1435 @fboehm
Thanks @csoneson and @vjcitn and @rcorty!

I believe that there was a temporary problem with the zenodo website. That should explain the difficulty in rendering the vignette that @vjcitn mentions.

@vjcitn, I was unable to reproduce the error from the tests about 'find_lin_indep_cols'. The tests succeeded when I ran devtools::test(). Should I be running tests with a different function call? I did see that I omitted a call to library(qtl2pleio) in the test file, test-drop_depcols.R. I fixed this omission. I was under the impression that loading packages (for use in testing) was needed only in the file testthat.R.

I'm unsure why R CMD check gives the NOTE that you mention. It seems that some of the packages that you list are, in fact, not used by my package. I have now deleted from the DESCRIPTION file's Imports field those packages that I don't use in qtl2pleio. However, some of the packages that your output lists are used in qtl2pleio: broman and devtools are two examples. I will look into this to try to understand if I've miscoded something.

I haven't yet undertaken the development of print methods for the objects that qtl2 and qtl2pleio functions create, but I think that it's a good idea. I will include them in a future version.

Regarding the definition of matrix x matrix and matrix x vector multiplications in C++ files, I took these lines of C++ code from the qtl2 package. I believe that existing C++ packages (libraries) define them in a slightly different fashion than what is used in my package. Specifically, my code gives errors when inputs' dimensions don't match. I don't think that the C++ libraries' default errors are as clear.

@rcorty, I haven't yet responded to your comments. I intend to do so very soon. Thanks again!
