### 2019-05-22 330feef
added call to library(qtl2pleio) to test file
