### 2019-05-22 #1435 @fboehm
@vjcitn, when I run R CMD check on the tar.gz file, I get a note like the one that you get, about the Imports field of the DESCRIPTION file. I have been adding to Imports field packages that I use in vignettes, even if I don't use them in my qtl2pleio functions. Should I not do this?
