### 2019-05-22 #1435 @vjcitn
On Wed, May 22, 2019 at 11:39 AM Frederick Boehm <notifications@github.com>
wrote:

> @vjcitn <https://github.com/vjcitn>, when I run R CMD check on the tar.gz
> file, I get a note like the one that you get, about the Imports field of
> the DESCRIPTION file. I have been adding to Imports field packages that I
> use in vignettes, even if I don't use them in my qtl2pleio functions.
> Should I not do this?
>

To my knowledge "Imports" concerns only code that will be compiled in the R
folder of the package.  Packages used only in
vignettes need not be loaded on package use, so can be kept out of
namespace processing.  Whether this is stated in WRE
or not I cannot say.  You could ask on r-devel mailing list for certainty.
