### 2019-05-22 #1435 @csoneson
>I have been adding to Imports field packages that I use in vignettes, even if I don't use them in my qtl2pleio functions. Should I not do this?

@fboehm The "R packages" book by Hadley Wickham contains useful information on this topic: http://r-pkgs.had.co.nz/description.html. Also WRE: https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Package-Dependencies
