### 2019-05-22 #1435 @vjcitn
from WRE: In particular, packages providing “only” data for examples or
vignettes should be listed in ‘Suggests’ rather than ‘Depends’ in order to
make lean installations possible.

On Wed, May 22, 2019 at 12:56 PM Charlotte Soneson <notifications@github.com>
wrote:

> I have been adding to Imports field packages that I use in vignettes, even
> if I don't use them in my qtl2pleio functions. Should I not do this?
>
> @fboehm <https://github.com/fboehm> The "R packages" book by Hadley
> Wickham contains useful information on this topic:
> http://r-pkgs.had.co.nz/description.html. Also WRE:
> https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Package-Dependencies
>
