+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: qtl2pleio
 Title: Testing pleiotropy vs. separate QTL in multiparental populations
-Version: 0.1.3
+Version: 0.1.3.9000
 Authors@R:
     person(given = "Frederick J",
            family = "Boehm",
@@ -26,12 +26,9 @@ Imports:
     ggplot2,
     magrittr,
     MASS,
-    microbenchmark,
     mvtnorm,
-    progress,
     Rcpp,
     RcppEigen,
-    readr,
     rlang,
     stringr,
     tibble,
