### 2019-05-23 1ffc395
fixed DESCRIPTION fields Imports and Suggests. Bumped version number
