### 2019-05-23 d51bcf1
switched to using zenodo to download scan results
