### 2019-05-23 3ecae43
added back RcppEigen in linkingto field of DESCRIPTION
