### 2019-06-01 #1435 @fboehm
Hi, @rcorty, I've tried to address the issues that you raise in the comments as I've made revisions to the package. Thank you!

@csoneson - what are the next steps in the review process? Thanks again!
