### 2019-06-01 9d3a01b
added "QTL" with goal of clarifying paper.md
