### 2019-06-01 #1435 @fboehm
Thanks, @csoneson!

In response to @rcorty's comments, I made the following changes:

1. added LICENSE.md file that contains the text of the MIT license.
2. I tweaked the wording of the sentence in paper.md. However, I avoided using the term "causal variant".
3. I didn't add precalculated genotype probabilities to the README. The reason for avoiding this is that the calculations in the README are done on only 3 chromosomes, and with a sparse marker set, so the calculations don't consume much computing time.

In response to @vjcitn's comments, I made these changes:

1. Fixed the DESCRIPTION file - specifically the Suggests and Imports fields - so that the package passes R CMD check. I was previously ignorant of the distinction between these two fields. @vjcitn's references to Writing R Extensions were particularly helpful here.
2. I added a line of R code - library(qtl2pleio) - to the specified test file to ensure that the test files can be run manually.
3. I haven't yet implemented print methods for the objects that I create in the README and vignettes. I did open an issue on my repo for this. Since many of the print methods would apply to objects that are outputted by the qtl2 R package's functions, I might work with its authors to see about adding print methods to qtl2. Still, print methods for outputs from qtl2pleio-specific functions would be useful, too.
