### 2019-06-10 #1435 @vjcitn
yes -- is it ready now

On Mon, Jun 10, 2019 at 3:04 PM Charlotte Soneson <notifications@github.com>
wrote:

> 👋@vjcitn <https://github.com/vjcitn>, @rcorty <https://github.com/rcorty>
> - do you think you will have the chance to look at the revised submission
> in the coming week?
>
