### 2019-06-18 #1435 @csoneson
ping @vjcitn, @rcorty
> 👋@vjcitn, @rcorty - do you think you will have the chance to look at the revised submission in the coming week?
