### 2019-06-18 #1435 @vjcitn
@csoneson  I have just looked at the paper which seems fine.  @fboehm I think it is somewhat problematic to have the recla.Rmd in the vignettes folder as it involves the user in a download of 250+MB from zenodo and will go slowly if one is running CMD check.  You could place that particular vignette in a dedicated folder in /inst.

Our dialogue on C++ does not need to go on indefinitely but I would comment that reproducing source code from qtl2 may not be the best strategy.  You are already using LinkingTo and you have qtl2 in Suggests.  It may be best to take whatever steps are necessary (perhaps moving qtl2 to Imports) to reuse the compiled code from qtl2 instead of replicating source.  This is a minor concern.

I consider these optional issues for the developer and am ready to endorse acceptance of this software.  @csoneson let me know what else I need to do, if anything.
