### 2019-06-18 #1435 @vjcitn
I have looked at the checklist and discovered that the References criterion on DOIs is not fulfilled.  For example, the Bates/Eddelbuettel paper has DOI http://dx.doi.org/10.18637/jss.v052.i05
