### 2019-06-18 #1435 @vjcitn
It is also not clear to me that "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" has been explicitly addressed.  The existence of Code of Conduct seems relevant, and presence in github implicitly addresses the concern.  A comment from the author on this matter in the manuscript or README.md could settle this.
