### 2019-06-18 #1435 @danielskatz
and it's not always correct - it's just suggestions for things a person needs to check
