### 2019-06-20 #1435 @csoneson
@fboehm - as noted by @vjcitn, please add DOIs to your references and add some guidelines for potential contributors and users looking for help, and we can tick off those boxes. I have emailed @rcorty and I hope he will also be able to take a look at the revised submission shortly.
