### 2019-06-20 555a6db
added dois to paper.bib; added contributor guidelines to README.Rmd; knitted README.Rmd to update README.md
