+++ paper.bib
@@ -6,7 +6,8 @@
   number={3},
   pages={1111--1127},
   year={1995},
-  publisher={Genetics Soc America}
+  publisher={Genetics Soc America},
+  doi={}
 }
 
 @article{knott2000multitrait,
@@ -39,7 +40,8 @@
   issue={1},
   pages={1-26},
   year={1979},
-  publisher={American Statistical Association}
+  publisher={American Statistical Association},
+  doi={doi:10.1214/aos/1176344552}
 }
 
 @article{tian2016dissection,
@@ -50,7 +52,8 @@
   number={4},
   pages={1563--1574},
   year={2016},
-  publisher={Genetics Soc America}
+  publisher={Genetics Soc America},
+  doi={https://doi.org/10.1534/genetics.115.183624}
 }
 
 @article{zhou2014efficient,
@@ -61,7 +64,8 @@
   number={4},
   pages={407},
   year={2014},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={https://doi.org/10.1038/nmeth.2848}
 }
 
 @article{kang2008efficient,
@@ -72,15 +76,16 @@
   number={3},
   pages={1709--1723},
   year={2008},
-  publisher={Genetics Soc America}
+  publisher={Genetics Soc America},
+  doi={https://doi.org/10.1534/genetics.107.080101}
 }
 
 @article{boehm2019testing,
   title={Testing pleiotropy vs. separate QTL in multiparental populations},
   author={Boehm, Frederick J. and Chesler, Elissa J. and Yandell, Brian S. and Broman, Karl W.},
-  journal={Biorxiv},
+  journal={G3: Genes, Genomes, Genetics},
   year={2019},
-  url={https://www.biorxiv.org/content/10.1101/550939v1}
+  doi={https://doi.org/10.1534/g3.119.400098}
   }
 
 
@@ -91,7 +96,8 @@
   volume={3},
   number={1},
   pages={5--10},
-  year={2011}
+  year={2011},
+  doi={https://doi.org/10.32614/rj-2011-002}
 }
 
 @article{eddelbuettel2011rcpp,
@@ -101,7 +107,8 @@
   volume={40},
   number={8},
   pages={1--18},
-  year={2011}
+  year={2011},
+  doi={https://doi.org/10.18637/jss.v052.i05}
 }
 
 @article{bates2013fast,
