### 2019-06-20 #1435 @fboehm
Thanks, @csoneson and @vjcitn! I just pushed to Github the updated bib file (with dois). I double-checked the dois for the articles that whedon identified. I also added contributor guidelines to README.md.
