+++ paper.bib
@@ -107,8 +107,7 @@
   volume={40},
   number={8},
   pages={1--18},
-  year={2011},
-  doi={https://doi.org/10.18637/jss.v052.i05}
+  year={2011}
 }
 
 @article{bates2013fast,
@@ -118,5 +117,6 @@
   volume={52},
   number={5},
   pages={1--24},
-  year={2013}
+  year={2013},
+  doi={https://doi.org/10.18637/jss.v052.i05}
 }
