### 2019-06-20 #1435 @csoneson
@fboehm - please double-check. I think you have added the RcppEigen DOI to the wrong bib entry.
