### 2019-06-20 cfc7134
added technical support section to README.Rmd
