### 2019-06-20 #1435 @csoneson
There are still a couple of missing DOIs. Please also check capitalization in the bibliography and add {} around letters if necessary.
