### 2019-06-20 #1435 @fboehm
Thanks - I had missed a few DOIs. Do all articles have DOIs? I'm unsure how to get the DOIs for the 1995 articles.
