### 2019-06-20 a1d2332
added doi for rcpp article
