+++ paper.bib
@@ -107,7 +107,8 @@
   volume={40},
   number={8},
   pages={1--18},
-  year={2011}
+  year={2011},
+  doi={http://dx.doi.org/10.18637/jss.v040.i08}
 }
 
 @article{bates2013fast,
