### 2019-06-20 e3db78b
tried to fix capitalization by adding braces around letters in bib file.
