+++ paper.bib
@@ -81,7 +81,7 @@
 }
 
 @article{boehm2019testing,
-  title={Testing pleiotropy vs. separate QTL in multiparental populations},
+  title={Testing pleiotropy vs. {s}eparate {QTL} in multiparental populations},
   author={Boehm, Frederick J. and Chesler, Elissa J. and Yandell, Brian S. and Broman, Karl W.},
   journal={G3: Genes, Genomes, Genetics},
   year={2019},
@@ -101,7 +101,7 @@
 }
 
 @article{eddelbuettel2011rcpp,
-  title={Rcpp: Seamless R and C++ integration},
+  title={Rcpp: Seamless {R} and {C}++ integration},
   author={Eddelbuettel, Dirk and Fran{\c{c}}ois, Romain and Allaire, J and Ushey, Kevin and Kou, Qiang and Russel, N and Chambers, John and Bates, D},
   journal={Journal of Statistical Software},
   volume={40},
@@ -112,7 +112,7 @@
 }
 
 @article{bates2013fast,
-  title={Fast and elegant numerical linear algebra using the RcppEigen package},
+  title={Fast and elegant numerical linear algebra using the {RcppEigen} package},
   author={Bates, Douglas and Eddelbuettel, Dirk},
   journal={Journal of Statistical Software},
   volume={52},
