### 2019-06-20 #1435 @fboehm
@csoneson - I now see the capitalization problems in the bibliography. Does adding braces around letters lead to no changes in the capitalization for those letters?
