### 2019-06-20 #1435 @csoneson
Yes - see e.g. https://texfaq.org/FAQ-capbibtex.
For the DOIs for the articles from 1995, I also can't find them (neither on PubMed nor on https://www.crossref.org/guestquery/). @openjournals/joss-eics - do you have additional suggestions where to look, or is it ok to leave it like this?
