### 2019-06-29 601bf34
added test for phenotypes matrix rank
