### 2019-06-29 #1435 @kyleniemeyer
Hi @fboehm, I made a few fixes to the bib file in https://github.com/fboehm/qtl2pleio/pull/20 (DOIs should not include the full URL in the bibtex file, just the DOI itself).
