### 2019-06-29 #1435 @csoneson
@kyleniemeyer Thanks for your input! Do you have a suggestion for where to possibly find DOIs for old articles, or is it ok to leave it without?
> For the DOIs for the articles from 1995, I also can't find them (neither on PubMed nor on https://www.crossref.org/guestquery/). @openjournals/joss-eics - do you have additional suggestions where to look, or is it ok to leave it like this?
