### 2019-06-29 #1435 @kyleniemeyer
@csoneson I went to the publisher's page for those articles, and they don't have one listed, so I don't think they have DOIs. Some publishers went back and generated DOIs for articles that were published prior to the adoption of DOIs, but apparently this one didn't. So I think it's fine—they just don't have DOIs.
