### 2019-06-29 #1435 @rcorty
I believe this article and package are passing all requirements and I've updated my check boxes to reflect this.
