### 2019-06-29 #1435 @vjcitn
I have updated check boxes.  Package/paper passes.
