### 2019-06-30 #1435 @csoneson
@fboehm - one more small thing to fix in the references:
* Efron 1979 - should be Annals of Statistics (with capital S)
I'm not quite sure why the Bates & Eddelbuettel reference shows up as "D. Bates & Eddelbuettel" in the text - if I compile the paper locally this doesn't happen, and the bib file looks fine.
