### 2019-06-30 f3a8463
fixed capitalization of AoS journal
