+++ paper.bib
@@ -35,7 +35,7 @@
 @article{efron1979bootstrap,
   title={Bootstrap methods: another look at the jackknife},
   author={Efron, Bradley},
-  journal={Annals of statistics},
+  journal={Annals of {S}tatistics},
   volume={7},
   issue={1},
   pages={1-26},
