### 2019-06-30 #1435 @fboehm
@csoneson - Thanks so much! I’ve fixed the bib file by making the first S a capital letter.
