### 2019-06-30 753a04f
bumped version number after adding .zenodo.json
