+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: qtl2pleio
 Title: Testing pleiotropy vs. separate QTL in multiparental populations
-Version: 1.0.0
+Version: 1.0.1
 Authors@R:
     person(given = "Frederick J",
            family = "Boehm",
