### 2019-06-30 f46c66d
fixed .zenodo.json and bumped version number
