### 2019-06-30 ce4adfc
bumped version number after another try at fixing .zenodo.json.
