### 2019-06-30 #1435 @csoneson
@openjournals/joss-eics - we are ready for acceptance here! The only weird thing is that the Bates & Eddelbuettel reference in the paper shows up as "D. Bates & Eddelbuettel" - however, the "D." is not included if I build the markdown locally.
