### 2019-06-30 #1435 @kyleniemeyer
@csoneson yeah, that's strange. Not sure why it's happening. Unless that's a major issue, I think we should just proceed.
