### 2019-06-30 #1435 @kyleniemeyer
@fboehm congrats on your submission's publication in JOSS! Thanks to @vjcitn and @rcorty for reviewing, and @csoneson for editing.
